fix: resolve compilation errors across multiple screens
- Replace driverVehiclesProvider with myVehiclesProvider in driver screens
- Add displayName getter to UserRole enum
- Fix IconButton? to Widget type error in user_details_screen
- Fix num to double type conversion in admin_repository

+++ lib/repositories/admin_repository.dart
@@ -94,8 +94,8 @@ class AdminRepository {
       tripsToday: tripsToday,
       tripsTodayChange: tripsTodayChange,
       revenueMtd: revenueMtd,
-      revenueMtdChange: revenueMtdChange,
-      fleetUtilization: fleetUtilization,
+      revenueMtdChange: revenueMtdChange.toDouble(),
+      fleetUtilization: fleetUtilization.toDouble(),
       activeVehicles: vehicleStats['active']!,
       totalVehicles: vehicleStats['total']!,
     );
